src/submission-intake/pyproject.toml
[project]
name = "submission-intake"
version = "0.1.0"
description = "Submission intake service for event sourcing workflow"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiohttp>=3.12.13",
    "azure-cosmos>=4.8.0",
    "azure-identity>=1.23.0",
    "azure-servicebus>=7.12.0",
    "pydantic>=2.10.0",
    "python-dotenv>=1.1.1",
]
